pyproject.toml
[tool.poetry]
name = "hugo-obsidian"
version = "0.1.0"
description = ""
authors = ["well"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
optfunc2 = "^0.2.6"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
